Coding_Teams-Autogen/requirements.txt
autogen-agentchat[anthropic]~=0.4
anthropic>=0.34.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.16.0
fastapi>=0.104.0
uvicorn>=0.24.0
rich>=13.0.0
typer>=0.12.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
mcp>=1.0.0
aiofiles>=23.0.0
watchdog>=3.0.0
pathspec>=0.11.0
statistics>=1.0.3.5
hashlib2>=1.0.0
difflib>=1.0.0
pickle5>=0.0.12 